refactor(core): split command and find boundaries
