+++ src/exomem/find.py
@@ -17,16 +17,21 @@ import re
 import threading
 import time
 from collections import OrderedDict
-from contextlib import contextmanager, nullcontext
 from dataclasses import dataclass, field
 from datetime import date, timedelta
-from functools import cached_property
 from pathlib import Path
 from typing import Any
 
 import yaml
 
 from . import freshness
+from .find_types import (
+    FindTimings,
+    Hit,
+    ParsedPage,
+    _format_timestamp as _format_timestamp,
+    timing_span as _span,
+)
 from .kbdir import kb_dirname, kb_prefix
 
 log = logging.getLogger(__name__)
@@ -103,330 +108,6 @@ def reset_active_ranking_cache() -> None:
 
 
 @dataclass
-class ParsedPage:
-    path: Path  # absolute
-    rel_path: str  # vault-relative, e.g. "Knowledge Base/Notes/Insights/foo.md"
-    frontmatter: dict[str, Any]
-    body: str
-    title: str
-    mtime: float
-
-    @property
-    def page_type(self) -> str | None:
-        t = self.frontmatter.get("type")
-        return str(t) if t else None
-
-    @property
-    def scope(self) -> str | None:
-        """Per-type "scope" field as defined in the plan.
-
-        - research-note → project
-        - pattern / insight / failure → project (singular) or projects (joined)
-        - experiment → domain
-        - production-log → medium
-        - entity → entity_type
-        - source → source_type
-        Fallback: project / projects / domain / medium / entity_type in that order.
-        """
-        fm = self.frontmatter
-        t = self.page_type
-
-        def _project_or_projects() -> str | None:
-            if (proj := fm.get("project")):
-                return str(proj)
-            if (projects := fm.get("projects")):
-                if isinstance(projects, list) and projects:
-                    return ",".join(str(p) for p in projects)
-                return str(projects)
-            return None
-
-        if t == "production-log":
-            return str(fm["medium"]) if fm.get("medium") else None
-        if t == "experiment":
-            return str(fm["domain"]) if fm.get("domain") else None
-        if t == "entity":
-            return str(fm["entity_type"]) if fm.get("entity_type") else None
-        if t == "source":
-            return str(fm["source_type"]) if fm.get("source_type") else None
-        if t in ("research-note", "pattern", "insight", "failure"):
-            return _project_or_projects()
-
-        # Unknown type: fall back across all candidates
-        return (
-            _project_or_projects()
-            or (str(fm["domain"]) if fm.get("domain") else None)
-            or (str(fm["medium"]) if fm.get("medium") else None)
-            or (str(fm["entity_type"]) if fm.get("entity_type") else None)
-        )
-
-    @property
-    def updated(self) -> str:
-        u = self.frontmatter.get("updated") or self.frontmatter.get("captured") or ""
-        return str(u)
-
-    @property
-    def tags(self) -> list[str]:
-        t = self.frontmatter.get("tags") or []
-        return [str(x).lower() for x in t] if isinstance(t, list) else []
-
-    @property
-    def speakers(self) -> list[str]:
-        """Named speakers from a diarized media sidecar's `speakers:` frontmatter
-        (e.g. `[Alice, Speaker B]`). Empty when undiarized or not a media sidecar."""
-        s = self.frontmatter.get("speakers") or []
-        return [str(x) for x in s] if isinstance(s, list) else []
-
-    @property
-    def media_type(self) -> str | None:
-        """audio/video/image/pdf on an Evidence media sidecar, else None."""
-        mt = self.frontmatter.get("media_type")
-        return str(mt) if mt else None
-
-    @property
-    def media_file(self) -> str | None:
-        """Vault-relative pointer to the original binary this sidecar describes."""
-        ef = self.frontmatter.get("evidence_file")
-        return str(ef) if ef else None
-
-    @property
-    def parent_media(self) -> str | None:
-        """Vault-relative path of the parent video for a scene-frame child sidecar
-        (written by scene_frames). None for every other page."""
-        pm = self.frontmatter.get("parent_media")
-        return str(pm) if pm else None
-
-    @property
-    def frame_ts(self) -> float | None:
-        """Seconds into the parent video for a scene-frame child (defensive parse)."""
-        v = self.frontmatter.get("frame_ts")
-        if v is None:
-            return None
-        try:
-            return float(v)
-        except (TypeError, ValueError):
-            return None
-
-    @property
-    def file_kind(self) -> str:
-        """Coarse artifact kind for file-type filtering: a dataset's underlying
-        format (csv/json/tsv), a binary companion's media_type (pdf/image/audio/
-        video), else 'note' for a plain markdown page. The vocabulary `find`'s
-        `file_types`/`exclude_file_types` scope on."""
-        if self.page_type == "dataset":
-            fmt = self.frontmatter.get("format")
-            return str(fmt).lower() if fmt else "dataset"
-        if self.media_type:
-            return self.media_type.lower()
-        return "note"
-
-    @property
-    def status(self) -> str | None:
-        """Lifecycle status — draft / active / superseded / archived (None if unset)."""
-        s = self.frontmatter.get("status")
-        return str(s) if s else None
-
-    @property
-    def superseded_by(self) -> list[str]:
-        """Wikilink(s) to the page(s) that replaced this one (empty when not superseded)."""
-        sb = self.frontmatter.get("superseded_by")
-        if not sb:
-            return []
-        return [str(x) for x in sb] if isinstance(sb, list) else [str(sb)]
-
-    @property
-    def supersedes(self) -> list[str]:
-        """Wikilink(s) to the page(s) this one replaced — the backward supersession
-        pointer `replace` writes on a new page (empty when it supersedes nothing)."""
-        sv = self.frontmatter.get("supersedes")
-        if not sv:
-            return []
-        return [str(x) for x in sv] if isinstance(sv, list) else [str(sv)]
-
-    # ---- Derived-text memoization (computed at most once per page revision;
-    # FrontmatterCache replaces the whole ParsedPage on mtime change, so these
-    # never go stale). Nothing query-dependent is cached here. ----
-
-    @cached_property
-    def body_stripped(self) -> str:
-        return self.body.strip()
-
-    @cached_property
-    def body_norm(self) -> str:
-        return self.body_stripped.lower()
-
-    @cached_property
-    def title_norm(self) -> str:
-        return self.title.lower()
-
-    @cached_property
-    def stem_set(self) -> frozenset[str]:
-        """Snowball stems of title+body — the corpus side of the stem gates."""
-        from . import bm25
-
-        return frozenset(bm25.tokenize(self.title + " " + self.body))
-
-
-def _format_timestamp(seconds: float) -> str:
-    """Seconds → `mm:ss` (or `h:mm:ss` past an hour) for human-readable video deeplinks."""
-    total = int(round(seconds))
-    h, rem = divmod(total, 3600)
-    m, s = divmod(rem, 60)
-    return f"{h}:{m:02d}:{s:02d}" if h else f"{m}:{s:02d}"
-
-
-@dataclass
-class Hit:
-    path: str
-    type: str | None
-    scope: str | None
-    title: str
-    updated: str
-    excerpt: str
-    # Per-mode ranking signals — populated by hybrid/vector modes only.
-    # `None` means the ranker did not surface this path in its top-K. The
-    # presence of these fields lets a caller introspect WHY a hit ranked
-    # (vector_rank=1 vs bm25_rank=1 vs graph_hop=True) without re-running
-    # the query in each mode. Always omitted from as_dict() when empty so
-    # keyword-mode callers don't see noise.
-    bm25_rank: int | None = None
-    vector_rank: int | None = None
-    vector_score: float | None = None
-    clip_rank: int | None = None      # rank from CLIP text→image visual search
-    clip_score: float | None = None   # CLIP cosine similarity (image vs the query)
-    graph_hop: bool = False
-    graph_in_degree: int = 0
-    keyword_rank: int | None = None
-    rerank_score: float | None = None
-    # True when this hit came from OUTSIDE Knowledge Base/ via scope="kb"
-    # auto-widening. Lets the caller (and SKILL.md guidance) see that the
-    # search reached past the curated KB into the wider vault. Omitted from
-    # as_dict() when False so KB-scoped callers don't see noise.
-    outside_kb: bool = False
-    # Set when the hit is an Evidence media sidecar — `media_type` is
-    # audio/video/image/pdf and `media_file` points at the original binary, so the
-    # caller surfaces the FILE as the result (and can mint_download_token it),
-    # with the matched transcript/OCR text as the "why". Omitted when absent.
-    media_type: str | None = None
-    media_file: str | None = None
-    # Seconds into a video where its best CLIP keyframe matched the query. Set only
-    # on video visual hits (multi-vector index); None for images. Surfaced as a
-    # human-readable `clip_match_at` ("14:32") so the caller can deep-link the moment.
-    clip_frame_ts: float | None = None
-    # Persisted scene frame backing this hit (vault-relative JPEG path) + its
-    # timestamp. Set when the hit matched via a scene frame's OCR text (frame
-    # children group under the parent video) or resolved as the nearest persisted
-    # frame to a CLIP keyframe match. The JPEG is directly downloadable — the
-    # visual "why" for the hit. Surfaced as `scene_frame` + `scene_match_at`.
-    scene_frame: str | None = None
-    scene_frame_ts: float | None = None
-    # Seconds into a TIMED transcript where the text match localizes — the
-    # matched semantic-segment chunk's leading [timestamp] (vector lane) or the
-    # nearest marker preceding the query anchor (BM25/keyword). Surfaced as
-    # `transcript_match_at`. Data-driven: flat sidecars never set it.
-    transcript_ts: float | None = None
-    # Lifecycle. `status` is set when a hit is NOT plain `active`, so a reader can
-    # tell a superseded tombstone (or draft) from a live conclusion; `superseded_by`
-    # carries the forward pointer(s) to the replacement(s). Surfaced in as_dict()
-    # regardless of `prefer_active` — exposure is independent of the ranking
-    # demotion. Omitted when status is active/unset so live hits stay noise-free.
-    status: str | None = None
-    superseded_by: list[str] = field(default_factory=list)
-    # Usage-activation transparency (prefer_used=true only): the raw ACT-R
-    # activation B and the multiplier it produced, so the caller sees exactly
-    # why and by how much a hit was boosted. None (omitted) whenever the
-    # boost didn't touch this hit — default responses are byte-identical.
-    activation: float | None = None
-    usage_boost_applied: float | None = None
-
-    def as_dict(self) -> dict:
-        out: dict = {
-            "path": self.path,
-            "type": self.type,
-            "scope": self.scope,
-            "title": self.title,
-            "updated": self.updated,
-            "excerpt": self.excerpt,
-        }
-        if self.media_type:
-            out["media_type"] = self.media_type
-        if self.media_file:
-            out["media_file"] = self.media_file
-        if self.clip_frame_ts is not None:
-            out["clip_match_at"] = _format_timestamp(self.clip_frame_ts)
-        if self.scene_frame:
-            out["scene_frame"] = self.scene_frame
-            if self.scene_frame_ts is not None:
-                out["scene_match_at"] = _format_timestamp(self.scene_frame_ts)
-        if self.transcript_ts is not None:
-            out["transcript_match_at"] = _format_timestamp(self.transcript_ts)
-        if self.outside_kb:
-            out["outside_kb"] = True
-        if self.status and self.status != "active":
-            out["status"] = self.status
-        if self.superseded_by:
-            out["superseded_by"] = self.superseded_by
-        signals: dict = {}
-        if self.bm25_rank is not None:
-            signals["bm25_rank"] = self.bm25_rank
-        if self.vector_rank is not None:
-            signals["vector_rank"] = self.vector_rank
-        if self.vector_score is not None:
-            signals["vector_score"] = round(self.vector_score, 4)
-        if self.clip_rank is not None:
-            signals["clip_rank"] = self.clip_rank
-        if self.clip_score is not None:
-            signals["clip_score"] = round(self.clip_score, 4)
-        if self.clip_frame_ts is not None:
-            signals["clip_frame_ts"] = round(self.clip_frame_ts, 2)
-        if self.graph_hop:
-            signals["graph_hop"] = True
-        if self.graph_in_degree:
-            signals["graph_in_degree"] = self.graph_in_degree
-        if self.keyword_rank is not None:
-            signals["keyword_rank"] = self.keyword_rank
-        if self.rerank_score is not None:
-            signals["rerank_score"] = round(self.rerank_score, 4)
-        if self.activation is not None:
-            signals["activation"] = round(self.activation, 4)
-        if self.usage_boost_applied is not None:
-            signals["usage_boost"] = round(self.usage_boost_applied, 4)
-        if signals:
-            out["signals"] = signals
-        return out
-
-    def as_compact_dict(self) -> dict:
-        """Token-cheap routing shape: enough metadata to pick a page for a
-        follow-up `get`, with the token-heavy `excerpt` and `signals` omitted
-        (rerun with detail="full" when you need the why)."""
-        out: dict = {
-            "path": self.path,
-            "type": self.type,
-            "scope": self.scope,
-            "title": self.title,
-            "updated": self.updated,
-        }
-        if self.media_type:
-            out["media_type"] = self.media_type
-        if self.media_file:
-            out["media_file"] = self.media_file
-        if self.clip_frame_ts is not None:
-            out["clip_match_at"] = _format_timestamp(self.clip_frame_ts)
-        if self.scene_frame:
-            out["scene_frame"] = self.scene_frame
-            if self.scene_frame_ts is not None:
-                out["scene_match_at"] = _format_timestamp(self.scene_frame_ts)
-        if self.transcript_ts is not None:
-            out["transcript_match_at"] = _format_timestamp(self.transcript_ts)
-        if self.outside_kb:
-            out["outside_kb"] = True
-        if self.status and self.status != "active":
-            out["status"] = self.status
-        if self.superseded_by:
-            out["superseded_by"] = self.superseded_by
-        return out
-
-
-@dataclass
 class FrontmatterCache:
     """Per-process cache of parsed pages, invalidated by mtime."""
 
@@ -448,50 +129,6 @@ class FrontmatterCache:
 
 
 _CACHE = FrontmatterCache()
-
-
-class FindTimings:
-    """Opt-in per-stage timing collector for one `find` call.
-
-    Records milliseconds per named retrieval stage plus total elapsed time and
-    hot-cache status. Carries NO note content — stage names, numbers, flags,
-    and short exception class names only (the spec forbids bodies, excerpts,
-    query-expanded text, and vectors in diagnostics).
-    """
-
-    def __init__(self) -> None:
-        self._t0 = time.perf_counter()
-        self.stages: dict[str, dict[str, Any]] = {}
-        self.cache: dict[str, Any] = {"enabled": False, "hit": False}
-        self.profile: dict[str, Any] = {}
-
-    @contextmanager
-    def span(self, name: str):
-        t0 = time.perf_counter()
-        try:
-            yield
-        finally:
-            entry = self.stages.setdefault(name, {})
-            entry["ms"] = round((time.perf_counter() - t0) * 1000.0, 3)
-
-    def skipped(self, name: str) -> None:
-        self.stages.setdefault(name, {})["skipped"] = True
-
-    def error(self, name: str, exc: BaseException) -> None:
-        self.stages.setdefault(name, {})["error"] = type(exc).__name__
-
-    def as_dict(self) -> dict[str, Any]:
-        return {
-            "total_ms": round((time.perf_counter() - self._t0) * 1000.0, 3),
-            "cache": dict(self.cache),
-            "profile": dict(self.profile),
-            "stages": {k: dict(v) for k, v in self.stages.items()},
-        }
-
-
-def _span(timings: FindTimings | None, name: str):
-    """A timing span when a collector is present, else a no-op context."""
-    return timings.span(name) if timings is not None else nullcontext()
 
 
 # --------------------------------------------------------------------------- #
